Update lmo-tippopenfile.php
PHP Warning:  A non-numeric value encountered in ../lmo/addon/tipp/lmo-tippopenfile.php on line 77

+++ lmo/addon/tipp/lmo-tippopenfile.php
@@ -71,10 +71,10 @@ for($i = 1; $i <= count($tippdaten); $i++) {
   $op2 = substr($dum[0], 0, 5);
   $op3 = substr($dum[0], 5)-1;
   if ($op3 == $st-1) {
-    ////////////////////////////////////////////////// nur der benötigte Spieltag wird eingelesen
+    ////////////////////////////////////////////////// nur der benÃ¶tigte Spieltag wird eingelesen
     $jksp = $dum[3];
     $op4 = substr($dum[1], 2)-1;
-    $op6 = substr($dum[1], 2, -1)-1;
+    $op6 = (int)substr($dum[1], 2, -1)-1;
     $op7 = substr($dum[1], -1)-1;
     $op8 = substr($dum[1], 0, 2);
     if ($lmtype == 0) {
@@ -124,7 +124,7 @@ for($i = 1; $i <= count($tippdaten); $i++) {
         }
       }
     } // ende else
-  } // ende nur benötigter st
+  } // ende nur benÃ¶tigter st
 } // ende for($i=1;$i<=count($tippdaten);$i++)
 clearstatcache();
